+++ apps/web/templates/modal/types.ts
@@ -1,4 +1,4 @@
-import { MyTemplateCardPropsInterface } from '@ui/card';
+import { MyTemplateCardPropsInterface } from 'ui';
 
 export interface MytemplateListModalInnerPropsInterface {
   templates: MyTemplateCardPropsInterface[];
