+++ packages/ui/button/index.ts
@@ -3,3 +3,4 @@ export { IconWithTextButton } from '@ui/button/IconWithText';
 export { FullWidthButton } from '@ui/button/FullWidth';
 export { IconButton } from '@ui/button/Icon';
 export { FreeSizeButton } from '@ui/button/FreeSizeButton';
+export type { DefaultButtonPropsInterface } from './types';
