+++ packages/ui/form/index.ts
@@ -1,1 +1,2 @@
 export { Form } from './Form';
+export type { FormPropsInterface } from './types';
