+++ packages/ui/toast/index.ts
@@ -1,3 +1,4 @@
 export { StyledToastBoxList } from './styles';
 export * from './Box';
 export * from './BoxListTop';
+export type { ToastBoxInterface } from './types';
